+++ shuup/admin/modules/products/forms/base_forms.py
@@ -54,6 +54,7 @@ class ProductBaseForm(MultiLanguageModelForm):
             "description",
             "keywords",
             "name",
+            "slug",
             "status_text",
             "variation_name",
         )